+++ package.json
@@ -26,6 +26,7 @@
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
     "express-handlebars": "^3.0.0",
+    "float-label-css": "^1.0.2",
     "knex": "^0.15.2",
     "method-override": "^3.0.0",
     "pg": "^7.4.3",
