+++ public/styles.css
@@ -1,5 +1,8 @@
+@import url(https://fonts.googleapis.com/css?family=Roboto|Unica+One);
 * {
-  margin: 0 0; }
+  margin: 0 0;
+  font-family: "Roboto", sans-serif;
+  box-sizing: border-box; }
 
 .wrapper {
   width: 100vw;
@@ -12,28 +15,49 @@
 .header {
   grid-column: 2/6;
   grid-row: 1/1;
+  margin: 60px 0px 60px 0px;
+  padding: 20px 0px 20px 0px;
+  border-top: 1px solid #d9d9d9;
   display: grid;
   grid-template-columns: repeat(6, 1fr);
   grid-column-gap: 10px; }
 
 .header-logo {
-  grid-column: 1/4; }
+  grid-column: 1/3;
+  padding: 25px 20px 20px 20px;
+  height: 50px; }
   .header-logo img {
     width: 100%;
     height: auto; }
 
 .header-about {
-  grid-column: 4/6; }
+  grid-column: 3/6;
+  padding: 20px;
+  border-left: 1px solid #d9d9d9;
+  border-right: 1px solid #d9d9d9;
+  font-size: 12px;
+  height: 50px; }
 
 .user-menu {
-  grid-column: 7/7; }
+  grid-column: 6/6;
+  padding: 20px;
+  text-align: center;
+  height: 25px; }
 
 .content {
   grid-column: 2/6;
   grid-row: 2/2; }
 
+.maincontent {
+  grid-column: 2/5; }
+
+.sidebar {
+  grid-column: 5/6; }
+
 .featured {
-  grid-column: 1/7; }
+  grid-column: 1/7;
+  padding-bottom: 30px;
+  border-bottom: 1px solid #d9d9d9; }
   .featured img {
     object-fit: cover;
     width: 100%;
@@ -42,19 +66,68 @@
 .gallery {
   grid-column: 2/6;
   grid-row: 1/1;
+  padding: 20px 0px 0px 0px;
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   grid-column-gap: 10px; }
+  .gallery a {
+    text-decoration: none;
+    color: #919195; }
+    .gallery a p {
+      padding-top: 10px; }
+  .gallery a:hover {
+    color: #418AB4; }
 
 .card {
   max-width: 100%;
-  grid-column: 1fr; }
+  grid-column: 1fr;
+  margin: 0px 0px 30px 0px; }
   .card img {
     width: 100%;
     height: auto; }
+
+.form {
+  max-width: 500px;
+  margin: 0 auto; }
+  .form label {
+    margin: 20px 0px 20px 0px;
+    color: #418AB4; }
+  .form input {
+    width: 100%; }
+  .form input[type=submit] {
+    background-color: #418AB4;
+    color: #ffffff;
+    padding: 15px;
+    font-size: 1.2em; }
+  .form input[type=submit]:hover {
+    background-color: #5D96B4; }
+  .form input[type=submit]:active {
+    background-color: #2A5C78;
+    outline: none; }
+
+.imagedetail {
+  max-width: 100%;
+  grid-column: 1fr;
+  margin: 0px 0px 30px 0px; }
+  .imagedetail img {
+    width: 100%;
+    height: auto; }
+
+.imagelist {
+  max-width: 100%;
+  grid-column: 1fr;
+  margin: 0px 0px 30px 0px; }
+  .imagelist img {
+    width: 100%;
+    height: auto; }
+  .imagelist p {
+    margin-bottom: 20px; }
 
 .footer {
   grid-column: 1/7;
   grid-row: 3/3;
   height: 80px;
-  background-color: aquamarine; }
+  background-color: #d9d9d9;
+  text-align: center;
+  padding-top: 40px;
+  margin-top: 80px; }
