+++ scss/styles.scss
@@ -1,6 +1,20 @@
+@import url(https://fonts.googleapis.com/css?family=Roboto|Unica+One);
+
+$headerfont: 'Unica One', cursive;
+$bodyfont: 'Roboto', sans-serif;
+
+$lightgrey: #d9d9d9;
+$medgrey: #919195;
+$blue: #418AB4;
+$bluelight: #5D96B4;
+$bluedark: #2A5C78;
+
 * {
     margin: 0 0;
+    font-family: $bodyfont;
+    box-sizing: border-box;
 };
+
 
 .wrapper {
     width: 100vw;
@@ -14,13 +28,18 @@
 .header {
     grid-column: 2/6;
     grid-row: 1/1;
+    margin: 60px 0px 60px 0px;
+    padding: 20px 0px 20px 0px;
+    border-top: 1px solid $lightgrey;
     display: grid;
     grid-template-columns: repeat(6, 1fr);
     grid-column-gap: 10px;
 };
 
 .header-logo {
-    grid-column: 1/4;
+    grid-column: 1/3;
+    padding: 25px 20px 20px 20px;
+    height: 50px;
     img {
         width: 100%;
         height: auto;
@@ -28,11 +47,19 @@
 }
 
 .header-about {
-    grid-column: 4/6;
+    grid-column: 3/6;
+    padding: 20px; 
+    border-left: 1px solid $lightgrey;
+    border-right: 1px solid $lightgrey;
+    font-size: 12px;
+    height: 50px;
 }
 
 .user-menu{
-    grid-column: 7/7;
+    grid-column: 6/6;
+    padding: 20px;
+    text-align: center;
+    height: 25px;
 }
 
 .content {
@@ -40,8 +67,18 @@
     grid-row: 2/2;
 }
 
+.maincontent {
+    grid-column: 2/5;
+}
+
+.sidebar {
+    grid-column: 5/6;
+}
+
 .featured {
     grid-column: 1/7;
+    padding-bottom: 30px;
+    border-bottom: 1px solid $lightgrey;
     img {
         object-fit: cover;
         width: 100%;
@@ -52,21 +89,82 @@
 .gallery {
     grid-column: 2/6;
     grid-row: 1/1;
+    padding: 20px 0px 0px 0px;
     display: grid;
     grid-template-columns: repeat(3, 1fr);
     grid-column-gap: 10px;
+    a {
+        text-decoration: none;
+        color: $medgrey; 
+        p {
+            padding-top: 10px;
+        }
+    };
+    a:hover {
+        color: $blue;
+    }
 };
 
 .card {
     max-width: 100%;
     grid-column: 1fr;
+    margin: 0px 0px 30px 0px;
     img {
         width: 100%;
         height: auto;
     }
 }
 
+.form {
+    max-width: 500px;
+    margin: 0 auto;
+    label {
+        margin: 20px 0px 20px 0px;
+        color: $blue;
+    }
+    
+    input {
+        width: 100%;
+    }
 
+    input[type=submit] {
+        background-color: $blue;
+        color: #ffffff;
+        padding: 15px;
+        font-size: 1.2em;
+    }
+    input[type=submit]:hover {
+        background-color: $bluelight;
+    }
+    input[type=submit]:active {
+        background-color: $bluedark;
+        outline: none;
+    }
+
+}
+
+.imagedetail {
+    max-width: 100%;
+    grid-column: 1fr;
+    margin: 0px 0px 30px 0px;
+    img {
+        width: 100%;
+        height: auto;
+    }
+}
+
+.imagelist {
+    max-width: 100%;
+    grid-column: 1fr;
+    margin: 0px 0px 30px 0px;
+    img {
+        width: 100%;
+        height: auto;
+    }
+    p {
+        margin-bottom: 20px;
+    }
+}
 
 
  
@@ -74,5 +172,8 @@
     grid-column: 1/7;
     grid-row: 3/3;
     height: 80px;
-    background-color: aquamarine;
+    background-color: $lightgrey;
+    text-align: center;
+    padding-top: 40px;
+    margin-top: 80px;
 }
