+++ views/edit.hbs
@@ -1,15 +1,21 @@
-<form action="/gallery/{{image_id}}?_method=PUT" method="post">
-  <label>Title:
-    <input type="text" name="title" value="{{title}}" id="" required>
-  </label>
-  <label>Author:
-    <input type="text" name="author" value="{{author}}" id="" required>
-  </label>
-  <label>Link:
-    <input type="text" name="link" value="{{link}}"id="" required>
-  </label>
-   <label>Description:
-    <textarea rows="10" cols="30" name="description" id="" required>{{description}}</textarea>
-  </label>
-  <input type="submit" value="Submit" required>
-</form>
+<div class='content'>
+    <form class='form' action="/{{image_id}}?_method=PUT" method="post">
+        <label class="has-float-label">
+          <input type="text" name="title" placeholder="Image title" value="{{title}}"required>
+          <span>Title</span>
+        </label>
+        <label class="has-float-label">
+          <input type="text" name="author" placeholder="Author name" value="{{author}}" required>
+          <span>Author</span>
+        </label>
+        <label class="has-float-label">
+          <input type="text" name="link" placeholder="URL" value="{{link}}" required>
+          <span>Image link</span>
+        </label>
+        <label class="has-float-label">
+          <textarea rows="10" cols="30" name="description" placeholder="Details about this image" id="" required>{{description}}</textarea>
+          <span>Description</span>
+        </label>
+        <input type="submit" value="Submit" required>
+    </form>
+</div>
