+++ views/home.hbs
@@ -2,7 +2,7 @@
     {{#each photos }}
         {{#if featured}}
             <div class='featured'>
-                <a href='/gallery/{{ image_id }}'>
+                <a href='/{{ image_id }}'>
                     <img src="{{ link }}" />    
                 </a>
             </div>
@@ -11,9 +11,11 @@
     {{/each}}
     <div class='gallery'>
         {{#each photos }}
-            <a class='card' href='/gallery/{{ image_id }}'>
+            <a class='card' href='/{{ image_id }}'>
                 <img class='card-image' src="{{ link }}" />
-                {{ title }}
+                <p> 
+                    {{ title }}
+                </p>
             </a>
             
         {{/each}}
@@ -23,5 +25,5 @@
        
 
 
-    <a class='button' href="/gallery/new">Upload</a>
+    <a class='button' href="/new">Upload</a>
 </div>
